requirements.txt
flask
gunicorn
requests
python-dotenv

# Fabric Scraping Module
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
asyncio

# AI & RAG Integration
openai==1.6.1
tiktoken==0.5.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4

# Image Processing
Pillow==10.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
tenacity==8.2.3
pydantic==2.5.3
python-slugify==8.0.1